added quit message build messages properly join line of signs are colored now update sign blockstates fix TabCompletes fix saving head set fix Lang colors fix players in game status if the game has ended fix deserialization of game locations fix hidden armorstands never getting loaded, if they aren't in the spawn chunks. Now we have to look them up every time we need them. show countdown as title when starting the game end game automatically if no players are left
next commit probably will completely remove armor stands and instead set the head-blocks directly.
That's because the entities are colored dark by the client because of block light 0 when the eyes are in a block. and sadly the eyes of a small armor stand is in its chest.
So the options to fix this would be package manipulation with ProtocolLib/MNS, using big armor stands (wich only fixes the problem if the feet can be in a well lit block) or going a simpler approach of not using armor stands.